refactor: decompose audio alignment runner into a dedicated package
