+++ src/frame_compare/alignment/__init__.py
@@ -1,0 +1,18 @@
+"""Audio alignment package."""
+
+from .core import apply_audio_alignment, resolve_alignment_reference
+from .formatting import format_alignment_output
+from .models import (
+    AudioAlignmentDisplayData,
+    AudioAlignmentSummary,
+    AudioMeasurementDetail,
+)
+
+__all__ = [
+    "AudioAlignmentDisplayData",
+    "AudioAlignmentSummary",
+    "AudioMeasurementDetail",
+    "apply_audio_alignment",
+    "format_alignment_output",
+    "resolve_alignment_reference",
+]
