+++ src/frame_compare/alignment/core.py
@@ -1,4 +1,4 @@
-"""Audio alignment orchestration helpers."""
+"""Core orchestration logic for audio alignment."""
 
 from __future__ import annotations
 
@@ -6,9 +6,7 @@ import logging
 import math
 import sys
 import time
-from collections.abc import Mapping as MappingABC
 from contextlib import nullcontext
-from dataclasses import dataclass, field
 from pathlib import Path
 from typing import (
     TYPE_CHECKING,
@@ -23,43 +21,29 @@ from typing import (
 )
 
 import click
-
 from src import audio_alignment
 from src.datatypes import AppConfig
-from src.frame_compare import vspreview
-from src.frame_compare.alignment_helpers import derive_frame_hint
-from src.frame_compare.cli_runtime import (
-    CLIAppError,
-    ClipPlan,
-    ensure_audio_alignment_block,
-)
+from src.frame_compare.cli_runtime import CLIAppError, ClipPlan, CliOutputManagerProtocol
 from src.frame_compare.config_helpers import coerce_config_flag as _coerce_config_flag
 from src.frame_compare.layout_utils import plan_label as _plan_label
 from src.frame_compare.metadata import match_override as _match_override
 
-
-def resolve_subdir(root: Path, relative: str, *, purpose: str, allow_absolute: bool = False) -> Path:
-    """Delegate to preflight.resolve_subdir without creating an import cycle."""
-
-    from src.frame_compare import preflight as _preflight
-
-    return _preflight.resolve_subdir(root, relative, purpose=purpose, allow_absolute=allow_absolute)
+from .models import (
+    AudioAlignmentDisplayData,
+    AudioAlignmentSummary,
+    AudioMeasurementDetail,
+)
+from .persistence import (
+    apply_suggestion_hints_to_details,
+    extract_suggestion_hints,
+    load_existing_entries,
+    resolve_subdir,
+)
 
 if TYPE_CHECKING:
     from src.audio_alignment import AlignmentMeasurement, AudioStreamInfo
-    from src.frame_compare.cli_runtime import CliOutputManagerProtocol, JsonTail
 
 logger = logging.getLogger(__name__)
-_render_vspreview_script = vspreview.render_script
-_persist_vspreview_script = vspreview.persist_script
-_write_vspreview_script = vspreview.write_script
-_prompt_vspreview_offsets = vspreview.prompt_offsets
-_apply_vspreview_manual_offsets = vspreview.apply_manual_offsets
-_launch_vspreview = vspreview.launch
-_format_vspreview_manual_command = vspreview.format_manual_command
-_resolve_vspreview_command = vspreview.resolve_command
-_VSPREVIEW_WINDOWS_INSTALL = vspreview.VSPREVIEW_WINDOWS_INSTALL
-_VSPREVIEW_POSIX_INSTALL = vspreview.VSPREVIEW_POSIX_INSTALL
 
 
 def _fps_to_float(value: tuple[int, int] | None) -> float:
@@ -71,22 +55,6 @@ def _fps_to_float(value: tuple[int, int] | None) -> float:
     if not denominator:
         return 0.0
     return float(numerator) / float(denominator)
-
-
-def _str_int_dict_factory() -> dict[str, int]:
-    return {}
-
-
-def _str_float_dict_factory() -> dict[str, float]:
-    return {}
-
-
-def _measurement_dict_factory() -> dict[str, "_AudioMeasurementDetail"]:
-    return {}
-
-
-def _str_list_factory() -> list[str]:
-    return []
 
 
 def _plan_fps_map(plans: Sequence[ClipPlan]) -> dict[Path, tuple[int, int]]:
@@ -108,72 +76,6 @@ def _plan_fps_map(plans: Sequence[ClipPlan]) -> dict[Path, tuple[int, int]]:
             fps_map[plan.path] = candidate
             break
     return fps_map
-
-
-@dataclass
-class _AudioAlignmentSummary:
-    """
-    Bundle of audio-alignment details used for reporting and persistence.
-    """
-
-    offsets_path: Path
-    reference_name: str
-    measurements: Sequence["AlignmentMeasurement"]
-    applied_frames: dict[str, int]
-    baseline_shift: int
-    statuses: dict[str, str]
-    reference_plan: ClipPlan
-    final_adjustments: dict[str, int]
-    swap_details: dict[str, str]
-    suggested_frames: dict[str, int] = field(default_factory=_str_int_dict_factory)
-    suggestion_mode: bool = False
-    manual_trim_starts: dict[str, int] = field(default_factory=_str_int_dict_factory)
-    vspreview_manual_offsets: dict[str, int] = field(default_factory=_str_int_dict_factory)
-    vspreview_manual_deltas: dict[str, int] = field(default_factory=_str_int_dict_factory)
-    measured_offsets: dict[str, "_AudioMeasurementDetail"] = field(default_factory=_measurement_dict_factory)
-
-
-@dataclass
-class _AudioMeasurementDetail:
-    """Snapshot of an audio alignment measurement for CLI/JSON reporting."""
-
-    label: str
-    stream: str
-    offset_seconds: Optional[float]
-    frames: Optional[int]
-    correlation: Optional[float]
-    status: str
-    applied: bool
-    note: Optional[str] = None
-
-
-@dataclass
-class _AudioAlignmentDisplayData:
-    """
-    Pre-rendered data used to present audio alignment results in the CLI.
-    """
-
-    stream_lines: list[str]
-    estimation_line: Optional[str]
-    offset_lines: list[str]
-    offsets_file_line: str
-    json_reference_stream: Optional[str]
-    json_target_streams: dict[str, str]
-    json_offsets_sec: dict[str, float]
-    json_offsets_frames: dict[str, int]
-    warnings: list[str]
-    preview_paths: list[str] = field(default_factory=_str_list_factory)
-    inspection_paths: list[str] = field(default_factory=_str_list_factory)
-    confirmation: Optional[str] = None
-    correlations: dict[str, float] = field(default_factory=_str_float_dict_factory)
-    threshold: float = 0.0
-    manual_trim_lines: list[str] = field(default_factory=_str_list_factory)
-    measurements: dict[str, _AudioMeasurementDetail] = field(default_factory=_measurement_dict_factory)
-
-
-AudioAlignmentSummary = _AudioAlignmentSummary
-AudioMeasurementDetail = _AudioMeasurementDetail
-AudioAlignmentDisplayData = _AudioAlignmentDisplayData
 
 
 def _resolve_alignment_reference(
@@ -206,7 +108,212 @@ def _resolve_alignment_reference(
     return plans[0]
 
 
-resolve_alignment_reference = _resolve_alignment_reference
+def resolve_alignment_reference(
+    plans: Sequence[ClipPlan],
+    analyze_path: Path,
+    reference_hint: str,
+) -> ClipPlan:
+    return _resolve_alignment_reference(plans, analyze_path, reference_hint)
+
+
+def _estimate_frames_from_seconds(
+    measurement: "AlignmentMeasurement",
+    plan_lookup: Mapping[str, ClipPlan],
+) -> Optional[int]:
+    """Derive a frame delta from the available FPS hints when frames are missing."""
+
+    if measurement.frames is not None:
+        return int(measurement.frames)
+
+    offset_seconds = float(measurement.offset_seconds or 0.0)
+    if math.isclose(offset_seconds, 0.0, abs_tol=1e-9):
+        return 0
+
+    def _from_float(fps_value: Optional[float]) -> Optional[int]:
+        if fps_value and fps_value > 0:
+            return int(round(offset_seconds * float(fps_value)))
+        return None
+
+    for candidate in (measurement.target_fps, measurement.reference_fps):
+        derived = _from_float(candidate)
+        if derived is not None:
+            return derived
+
+    plan = plan_lookup.get(measurement.file.name)
+    if plan is not None:
+        fps_candidates = (
+            plan.effective_fps,
+            plan.applied_fps,
+            plan.fps_override,
+            plan.source_fps,
+        )
+        for fps_tuple in fps_candidates:
+            fps_float = _fps_to_float(fps_tuple)
+            derived = _from_float(fps_float)
+            if derived is not None:
+                return derived
+    return None
+
+
+def apply_manual_offsets_logic(
+    plans: Sequence[ClipPlan],
+    vspreview_reuse: dict[str, int],
+    display_data: AudioAlignmentDisplayData,
+    plan_labels: dict[Path, str],
+) -> tuple[dict[str, int], dict[str, int]]:
+    """
+    Apply manual offsets from VSPreview history, normalizing to avoid negative trims (padding).
+    Extracted for testability.
+
+    Returns:
+        tuple[dict[str, int], dict[str, int]]: (delta_map, manual_trim_starts)
+    """
+    if display_data.manual_trim_lines:
+        display_data.manual_trim_lines.clear()
+
+    baseline_map = {plan.path.name: int(plan.trim_start) for plan in plans}
+    delta_map: dict[str, int] = {}
+    manual_trim_starts: dict[str, int] = {}
+
+    # 1. Calculate proposed trims for all plans
+    proposed_trims: dict[str, int] = {}
+    for plan in plans:
+        key = plan.path.name
+        baseline = baseline_map.get(key, int(plan.trim_start))
+        delta = int(vspreview_reuse.get(key, 0))
+        proposed_trims[key] = baseline + delta
+        if key in vspreview_reuse:
+            delta_map[key] = delta
+
+    logger.info(f"[ALIGN DEBUG] Raw manual offsets (vspreview_reuse): {vspreview_reuse}")
+    logger.info(f"[ALIGN DEBUG] Baseline trims: {baseline_map}")
+    logger.info(f"[ALIGN DEBUG] Proposed trims before normalization: {proposed_trims}")
+
+    # 2. Normalize to ensure no negative trims (avoid padding)
+    # Normalize by shifting all clips up so the minimum trim becomes 0
+    # Example: offsets [-361, 0] → shift +361 → trims [0, 361]
+    min_offset = min(proposed_trims.values()) if proposed_trims else 0
+    shift = -min_offset if min_offset < 0 else 0
+
+    if shift > 0:
+        logger.info(f"[ALIGN DEBUG] Found negative proposed trim {min_offset}, applying global shift of {shift}f.")
+    else:
+        logger.info("[ALIGN DEBUG] No negative proposed trims found, no global shift applied.")
+
+    logger.info(f"[ALIGN DEBUG] Normalizing with shift={shift}")
+
+    # 3. Apply normalized trims
+    for plan in plans:
+        key = plan.path.name
+        if key in proposed_trims:
+            raw_offset = proposed_trims[key]
+            # Calculate normalized trim
+            normalized_trim = raw_offset + shift
+
+            # Apply to plan
+            # Note: We SET the new absolute trim_start (normalized_trim includes the baseline)
+            original_trim = int(plan.trim_start)
+            if normalized_trim != original_trim:
+                plan.trim_start = normalized_trim
+                plan.source_num_frames = None
+                plan.has_trim_start_override = True
+
+            manual_trim_starts[key] = int(plan.trim_start)
+
+            baseline = baseline_map.get(key, 0)
+            effective_delta = normalized_trim - baseline
+            delta_map[key] = effective_delta
+
+            label = plan_labels.get(plan.path, key)
+            input_delta = proposed_trims[key] - baseline
+
+            line = (
+                f"VSPreview manual offset applied: {label} baseline {baseline}f "
+                f"{input_delta:+d}f (shift {shift:+d}f) → {normalized_trim}f"
+            )
+            display_data.manual_trim_lines.append(line)
+
+            logger.info(f"[ALIGN DEBUG] Applied trim to {key}: raw={raw_offset} + shift={shift} -> {normalized_trim}. Final trim_start={plan.trim_start}")
+
+    return delta_map, manual_trim_starts
+
+
+def _reuse_vspreview_manual_offsets_if_available(
+    plans: Sequence[ClipPlan],
+    reference: ClipPlan | None,
+    vspreview_enabled: bool,
+    audio_cfg: Any,
+    display_data: AudioAlignmentDisplayData,
+    plan_labels: Dict[Path, str],
+    offsets_path: Path,
+) -> AudioAlignmentSummary | None:
+    if not (vspreview_enabled and reference and plans):
+        return None
+
+    try:
+        _, existing_entries = load_existing_entries(offsets_path)
+    except CLIAppError:
+        if not audio_cfg.enable:
+            return None
+        raise
+
+    vspreview_reuse: Dict[str, int] = {}
+    allowed_keys = {plan.path.name for plan in plans}
+    for key, value in existing_entries.items():
+        entry = value
+        status_obj = entry.get("status")
+        note_obj = entry.get("note")
+        frames_obj = entry.get("frames")
+        if not isinstance(status_obj, str) or not isinstance(frames_obj, (int, float)):
+            continue
+        if status_obj.strip().lower() != "manual":
+            continue
+        note_text = str(note_obj or "").strip().lower()
+        if "vspreview" not in note_text:
+            continue
+        if key in allowed_keys:
+            vspreview_reuse[key] = int(frames_obj)
+
+    if not vspreview_reuse:
+        return None
+
+    # Logic extracted to apply_manual_offsets_logic for testability
+    delta_map, manual_trim_starts = apply_manual_offsets_logic(
+        plans=plans,
+        vspreview_reuse=vspreview_reuse,
+        display_data=display_data,
+        plan_labels=plan_labels,
+    )
+
+    if not display_data.manual_trim_lines:
+        return None
+
+    display_data.offset_lines = ["Audio offsets: VSPreview manual offsets applied"]
+    if display_data.manual_trim_lines:
+        display_data.offset_lines.extend(display_data.manual_trim_lines)
+
+    label_map = {plan.path.name: plan_labels.get(plan.path, plan.path.name) for plan in plans}
+    display_data.json_offsets_frames = {
+        label_map.get(key, key): int(value)
+        for key, value in delta_map.items()
+    }
+    statuses_map = {key: "manual" for key in delta_map}
+    return AudioAlignmentSummary(
+        offsets_path=offsets_path,
+        reference_name=reference.path.name,
+        measurements=(),
+        applied_frames=dict(delta_map),
+        baseline_shift=0,
+        statuses=statuses_map,
+        reference_plan=reference,
+        final_adjustments=dict(manual_trim_starts),
+        swap_details={},
+        suggested_frames={},
+        suggestion_mode=False,
+        manual_trim_starts=manual_trim_starts,
+        vspreview_manual_offsets=dict(manual_trim_starts),
+        vspreview_manual_deltas=dict(delta_map),
+    )
 
 
 def apply_audio_alignment(
@@ -216,7 +323,7 @@ def apply_audio_alignment(
     root: Path,
     audio_track_overrides: Mapping[str, int],
     reporter: CliOutputManagerProtocol | None = None,
-) -> tuple[_AudioAlignmentSummary | None, _AudioAlignmentDisplayData | None]:
+) -> tuple[AudioAlignmentSummary | None, AudioAlignmentDisplayData | None]:
     """Apply audio alignment when enabled, returning summary and display data."""
     audio_cfg = cfg.audio_alignment
     prompt_reuse_offsets = _coerce_config_flag(audio_cfg.prompt_reuse_offsets)
@@ -225,7 +332,7 @@ def apply_audio_alignment(
         audio_cfg.offsets_filename,
         purpose="audio_alignment.offsets_filename",
     )
-    display_data = _AudioAlignmentDisplayData(
+    display_data = AudioAlignmentDisplayData(
         stream_lines=[],
         estimation_line=None,
         offset_lines=[],
@@ -253,151 +360,15 @@ def apply_audio_alignment(
     name_to_label: Dict[str, str] = {plan.path.name: plan_labels[plan.path] for plan in plans}
     plan_fps_map = _plan_fps_map(plans)
 
-    existing_entries_cache: tuple[str | None, Dict[str, Mapping[str, Any]]] | None = None
-
-    def _safe_float(value: object) -> float | None:
-        if isinstance(value, (int, float)):
-            float_value = float(value)
-            if math.isnan(float_value):
-                return None
-            return float_value
-        return None
-
-    def _safe_int(value: object) -> int | None:
-        if isinstance(value, (int, float)):
-            float_value = float(value)
-            if math.isnan(float_value):
-                return None
-            return int(float_value)
-        return None
-
-    def _extract_suggestion_hints(
-        entries: Mapping[str, Mapping[str, Any]],
-    ) -> Dict[str, tuple[int | None, float | None]]:
-        hints: Dict[str, tuple[int | None, float | None]] = {}
-        for name, entry in entries.items():
-            frames_hint = _safe_int(entry.get("suggested_frames"))
-            seconds_hint = _safe_float(entry.get("suggested_seconds"))
-            if frames_hint is None and seconds_hint is None:
-                continue
-            hints[name] = (frames_hint, seconds_hint)
-        return hints
-
-    def _apply_suggestion_hints_to_details(
-        detail_map: Dict[str, _AudioMeasurementDetail],
-        hints: Mapping[str, tuple[int | None, float | None]],
-    ) -> None:
-        for clip_name, (frames_hint, seconds_hint) in hints.items():
-            detail = detail_map.get(clip_name)
-            if detail is None:
-                continue
-            if frames_hint is not None:
-                detail.frames = int(frames_hint)
-            if seconds_hint is not None:
-                detail.offset_seconds = float(seconds_hint)
-
-    def _load_existing_entries() -> tuple[str | None, Dict[str, Mapping[str, Any]]]:
-        nonlocal existing_entries_cache
-        if existing_entries_cache is None:
-            try:
-                reference_name, existing_entries_raw = audio_alignment.load_offsets(
-                    offsets_path
-                )
-            except audio_alignment.AudioAlignmentError as exc:
-                raise CLIAppError(
-                    f"Failed to read audio offsets file: {exc}",
-                    rich_message=f"[red]Failed to read audio offsets file:[/red] {exc}",
-                ) from exc
-            normalized_entries: Dict[str, Mapping[str, Any]] = {}
-            if isinstance(existing_entries_raw, MappingABC):
-                raw_entries = cast(Mapping[Any, Any], existing_entries_raw)
-                for key_obj, value in raw_entries.items():
-                    if not isinstance(value, MappingABC):
-                        continue
-                    entry_mapping = cast(Mapping[Any, Any], value)
-                    normalized_entries[str(key_obj)] = {
-                        str(sub_key): sub_value for sub_key, sub_value in entry_mapping.items()
-                    }
-
-            existing_entries_cache = (
-                reference_name,
-                normalized_entries,
-            )
-        return existing_entries_cache
-
-    def _reuse_vspreview_manual_offsets_if_available(
-        reference: ClipPlan | None,
-    ) -> _AudioAlignmentSummary | None:
-        if not (vspreview_enabled and reference and plans):
-            return None
-
-        try:
-            _, existing_entries = _load_existing_entries()
-        except CLIAppError:
-            if not audio_cfg.enable:
-                return None
-            raise
-
-        vspreview_reuse: Dict[str, int] = {}
-        allowed_keys = {plan.path.name for plan in plans}
-        for key, value in existing_entries.items():
-            entry = value
-            status_obj = entry.get("status")
-            note_obj = entry.get("note")
-            frames_obj = entry.get("frames")
-            if not isinstance(status_obj, str) or not isinstance(frames_obj, (int, float)):
-                continue
-            if status_obj.strip().lower() != "manual":
-                continue
-            note_text = str(note_obj or "").strip().lower()
-            if "vspreview" not in note_text:
-                continue
-            if key in allowed_keys:
-                vspreview_reuse[key] = int(frames_obj)
-
-        if not vspreview_reuse:
-            return None
-
-        # Logic extracted to apply_manual_offsets_logic for testability
-        delta_map, manual_trim_starts = apply_manual_offsets_logic(
-            plans=plans,
-            vspreview_reuse=vspreview_reuse,
-            display_data=display_data,
-            plan_labels=plan_labels,
-        )
-
-        if not display_data.manual_trim_lines:
-            return None
-
-
-        display_data.offset_lines = ["Audio offsets: VSPreview manual offsets applied"]
-        if display_data.manual_trim_lines:
-            display_data.offset_lines.extend(display_data.manual_trim_lines)
-
-        label_map = {plan.path.name: plan_labels.get(plan.path, plan.path.name) for plan in plans}
-        display_data.json_offsets_frames = {
-            label_map.get(key, key): int(value)
-            for key, value in delta_map.items()
-        }
-        statuses_map = {key: "manual" for key in delta_map}
-        return _AudioAlignmentSummary(
-            offsets_path=offsets_path,
-            reference_name=reference.path.name,
-            measurements=(),
-            applied_frames=dict(delta_map),
-            baseline_shift=0,
-            statuses=statuses_map,
-            reference_plan=reference,
-            final_adjustments=dict(manual_trim_starts),
-            swap_details={},
-            suggested_frames={},
-            suggestion_mode=False,
-            manual_trim_starts=manual_trim_starts,
-            vspreview_manual_offsets=dict(manual_trim_starts),
-            vspreview_manual_deltas=dict(delta_map),
-        )
-
-    reused_summary = _reuse_vspreview_manual_offsets_if_available(reference_plan)
+    reused_summary = _reuse_vspreview_manual_offsets_if_available(
+        plans,
+        reference_plan,
+        vspreview_enabled,
+        audio_cfg,
+        display_data,
+        plan_labels,
+        offsets_path,
+    )
     if reused_summary is not None:
         if not audio_cfg.enable:
             display_data.warnings.append(
@@ -424,7 +395,7 @@ def apply_audio_alignment(
             display_data.warnings.append(
                 "[AUDIO] VSPreview manual alignment enabled — audio alignment disabled."
             )
-            summary = _AudioAlignmentSummary(
+            summary = AudioAlignmentSummary(
                 offsets_path=offsets_path,
                 reference_name=reference_plan.path.name,
                 measurements=(),
@@ -453,54 +424,16 @@ def apply_audio_alignment(
     measurement_order = [plan.path.name for plan in plans]
     negative_override_notes: Dict[str, str] = {}
 
-    def _estimate_frames_from_seconds(
-        measurement: "AlignmentMeasurement",
-        plan_lookup: Mapping[str, ClipPlan],
-    ) -> Optional[int]:
-        """Derive a frame delta from the available FPS hints when frames are missing."""
-
-        if measurement.frames is not None:
-            return int(measurement.frames)
-
-        offset_seconds = float(measurement.offset_seconds or 0.0)
-        if math.isclose(offset_seconds, 0.0, abs_tol=1e-9):
-            return 0
-
-        def _from_float(fps_value: Optional[float]) -> Optional[int]:
-            if fps_value and fps_value > 0:
-                return int(round(offset_seconds * float(fps_value)))
-            return None
-
-        for candidate in (measurement.target_fps, measurement.reference_fps):
-            derived = _from_float(candidate)
-            if derived is not None:
-                return derived
-
-        plan = plan_lookup.get(measurement.file.name)
-        if plan is not None:
-            fps_candidates = (
-                plan.effective_fps,
-                plan.applied_fps,
-                plan.fps_override,
-                plan.source_fps,
-            )
-            for fps_tuple in fps_candidates:
-                fps_float = _fps_to_float(fps_tuple)
-                derived = _from_float(fps_float)
-                if derived is not None:
-                    return derived
-        return None
-
     def _maybe_reuse_cached_offsets(
         reference: ClipPlan,
         candidate_targets: Sequence[ClipPlan],
-    ) -> _AudioAlignmentSummary | None:
+    ) -> AudioAlignmentSummary | None:
         if not prompt_reuse_offsets:
             return None
         if not sys.stdin.isatty():
             return None
         try:
-            cached_reference, existing_entries = _load_existing_entries()
+            cached_reference, existing_entries = load_existing_entries(offsets_path)
         except CLIAppError:
             return None
         if not existing_entries:
@@ -778,7 +711,7 @@ def apply_audio_alignment(
                 if plan is reference:
                     plan.alignment_status = "baseline"
 
-        summary = _AudioAlignmentSummary(
+        summary = AudioAlignmentSummary(
             offsets_path=offsets_path,
             reference_name=reference.path.name,
             measurements=measurements,
@@ -946,7 +879,7 @@ def apply_audio_alignment(
         else:
             display_data.stream_lines.append(f"Audio streams: target={target_stream_text}")
 
-    def _format_measurement_line(detail: _AudioMeasurementDetail) -> str:
+    def _format_measurement_line(detail: AudioMeasurementDetail) -> str:
         stream_text = detail.stream or "?"
         seconds_text = (
             f"{detail.offset_seconds:+.3f}s"
@@ -973,7 +906,7 @@ def apply_audio_alignment(
         )
 
     def _emit_measurement_lines(
-        detail_map: Dict[str, _AudioMeasurementDetail],
+        detail_map: Dict[str, AudioMeasurementDetail],
         order: Sequence[str],
         *,
         append_manual: bool = False,
@@ -1027,7 +960,7 @@ def apply_audio_alignment(
         manual_trims: Mapping[str, int],
         swap_map: Mapping[str, str],
         negative_notes: Mapping[str, str],
-    ) -> Dict[str, _AudioMeasurementDetail]:
+    ) -> Dict[str, AudioMeasurementDetail]:
         """
         Convert raw measurement objects into detail records used for CLI + JSON reporting.
 
@@ -1041,10 +974,10 @@ def apply_audio_alignment(
             negative_notes: Notes produced when negative offsets were redirected.
 
         Returns:
-            Dict[str, _AudioMeasurementDetail]: Mapping keyed by clip filename.
+            Dict[str, AudioMeasurementDetail]: Mapping keyed by clip filename.
         """
 
-        detail_map: Dict[str, _AudioMeasurementDetail] = {}
+        detail_map: Dict[str, AudioMeasurementDetail] = {}
         for measurement in measurement_seq:
             clip_name = measurement.file.name
             label = name_to_label.get(clip_name, clip_name)
@@ -1079,7 +1012,7 @@ def apply_audio_alignment(
                     status_text = "error"
                 applied_flag = False
             note_value = " ".join(note_parts) if note_parts else None
-            detail_map[clip_name] = _AudioMeasurementDetail(
+            detail_map[clip_name] = AudioMeasurementDetail(
                 label=label,
                 stream=descriptor,
                 offset_seconds=seconds_value,
@@ -1097,7 +1030,7 @@ def apply_audio_alignment(
             descriptor = stream_descriptors.get(clip_name, "")
             fps_value = fps_lookup.get(clip_name, 0.0)
             seconds_value = (trim_frames / fps_value) if fps_value else None
-            detail_map[clip_name] = _AudioMeasurementDetail(
+            detail_map[clip_name] = AudioMeasurementDetail(
                 label=label,
                 stream=descriptor,
                 offset_seconds=seconds_value,
@@ -1113,11 +1046,11 @@ def apply_audio_alignment(
     if reused_cached is not None:
         summary = reused_cached
         try:
-            _, existing_entries = _load_existing_entries()
+            _, existing_entries = load_existing_entries(offsets_path)
         except CLIAppError:
             existing_entries = {}
-        suggestion_hints = _extract_suggestion_hints(existing_entries)
-        detail_map: Dict[str, _AudioMeasurementDetail] = {}
+        suggestion_hints = extract_suggestion_hints(existing_entries)
+        detail_map: Dict[str, AudioMeasurementDetail] = {}
         for plan in plans:
             key = plan.path.name
             frames_val = summary.applied_frames.get(key)
@@ -1128,7 +1061,7 @@ def apply_audio_alignment(
             else:
                 seconds_val = frames_val / fps_val if fps_val else None
             descriptor = stream_descriptors.get(key, "")
-            detail_map[key] = _AudioMeasurementDetail(
+            detail_map[key] = AudioMeasurementDetail(
                 label=name_to_label.get(key, key),
                 stream=descriptor,
                 offset_seconds=seconds_val,
@@ -1137,7 +1070,7 @@ def apply_audio_alignment(
                 status=summary.statuses.get(key, "manual"),
                 applied=True,
             )
-        _apply_suggestion_hints_to_details(detail_map, suggestion_hints)
+        apply_suggestion_hints_to_details(detail_map, suggestion_hints)
         for name, (frames_hint, _seconds_hint) in suggestion_hints.items():
             if frames_hint is None:
                 continue
@@ -1287,8 +1220,6 @@ def apply_audio_alignment(
         for warning_message in dict.fromkeys(raw_warning_messages):
             _warn(warning_message)
 
-
-
         manual_trim_starts: Dict[str, int] = {}
         if vspreview_enabled:
             for plan in plans:
@@ -1301,7 +1232,7 @@ def apply_audio_alignment(
             display_data.warnings.append(
                 "[AUDIO] VSPreview manual alignment enabled — offsets reported for guidance only."
             )
-            summary = _AudioAlignmentSummary(
+            summary = AudioAlignmentSummary(
                 offsets_path=offsets_path,
                 reference_name=reference_plan.path.name,
                 measurements=measurements,
@@ -1336,7 +1267,7 @@ def apply_audio_alignment(
             offsets_path,
             reference_plan.path.name,
             measurements,
-            _load_existing_entries()[1],
+            load_existing_entries(offsets_path)[1],
             negative_override_notes,
         )
 
@@ -1369,7 +1300,7 @@ def apply_audio_alignment(
                 if plan is reference_plan:
                     plan.alignment_status = "baseline"
 
-        summary = _AudioAlignmentSummary(
+        summary = AudioAlignmentSummary(
             offsets_path=offsets_path,
             reference_name=reference_plan.path.name,
             measurements=measurements,
@@ -1406,251 +1337,19 @@ def apply_audio_alignment(
         ) from exc
 
 
-def format_alignment_output(
-    plans: Sequence[ClipPlan],
-    summary: _AudioAlignmentSummary | None,
-    display: _AudioAlignmentDisplayData | None,
-    *,
-    cfg: AppConfig,
-    root: Path,
-    reporter: CliOutputManagerProtocol,
-    json_tail: JsonTail,
-    vspreview_mode: str,
-    collected_warnings: List[str] | None = None,
-) -> None:
-    """
-    Populate json_tail/audio layout data and optionally launch VSPreview.
-    """
-
-    audio_block = ensure_audio_alignment_block(json_tail)
-
-    vspreview_target_plan: ClipPlan | None = None
-    vspreview_suggested_frames_value: int | None = None
-    vspreview_suggested_seconds_value = 0.0
-    if summary is not None:
-        for plan in plans:
-            if plan is summary.reference_plan:
-                continue
-            vspreview_target_plan = plan
-            break
-        if vspreview_target_plan is not None:
-            clip_key = vspreview_target_plan.path.name
-            vspreview_suggested_frames_value = derive_frame_hint(summary, clip_key)
-            measurement_seconds: Optional[float] = None
-            if summary.measured_offsets:
-                detail = summary.measured_offsets.get(clip_key)
-                if detail and detail.offset_seconds is not None:
-                    measurement_seconds = float(detail.offset_seconds)
-            if measurement_seconds is None and summary.measurements:
-                measurement_lookup = {
-                    measurement.file.name: measurement for measurement in summary.measurements
-                }
-                measurement = measurement_lookup.get(clip_key)
-                if measurement is not None and measurement.offset_seconds is not None:
-                    measurement_seconds = float(measurement.offset_seconds)
-            if measurement_seconds is not None:
-                vspreview_suggested_seconds_value = measurement_seconds
-
-    json_tail["vspreview_mode"] = vspreview_mode
-    json_tail["suggested_frames"] = (
-        int(vspreview_suggested_frames_value)
-        if vspreview_suggested_frames_value is not None
-        else None
-    )
-    json_tail["suggested_seconds"] = float(round(vspreview_suggested_seconds_value, 6))
-
-    vspreview_enabled_for_session = _coerce_config_flag(cfg.audio_alignment.use_vspreview)
-    if vspreview_enabled_for_session and summary is not None and summary.suggestion_mode:
-        try:
-            _launch_vspreview(
-                plans,
-                summary,
-                display,
-                cfg,
-                root,
-                reporter,
-                json_tail,
-            )
-        except CLIAppError:
-            raise
-        except Exception as exc:
-            logger.warning(
-                "VSPreview launch failed: %s",
-                exc,
-                exc_info=logger.isEnabledFor(logging.DEBUG),
-            )
-            reporter.warn(f"VSPreview launch failed: {exc}")
-
-    if display is not None:
-        audio_block["offsets_filename"] = display.offsets_file_line.split(": ", 1)[-1]
-        audio_block["reference_stream"] = display.json_reference_stream
-        target_streams: dict[str, object] = dict(display.json_target_streams.items())
-        audio_block["target_stream"] = target_streams
-        offsets_sec_source = dict(display.json_offsets_sec)
-        offsets_frames_source = dict(display.json_offsets_frames)
-        if (
-            not offsets_sec_source
-            and summary is not None
-            and summary.measured_offsets
-        ):
-            offsets_sec_source = {}
-            offsets_frames_source = {}
-            for detail in summary.measured_offsets.values():
-                if detail.offset_seconds is not None:
-                    offsets_sec_source[detail.label] = float(detail.offset_seconds)
-                if detail.frames is not None:
-                    offsets_frames_source[detail.label] = int(detail.frames)
-        audio_block["offsets_sec"] = {key: float(value) for key, value in offsets_sec_source.items()}
-        audio_block["offsets_frames"] = {
-            key: int(value) for key, value in offsets_frames_source.items()
-        }
-        stream_lines_output = list(display.stream_lines)
-        if display.estimation_line:
-            stream_lines_output.append(display.estimation_line)
-        audio_block["stream_lines"] = stream_lines_output
-        audio_block["stream_lines_text"] = "\n".join(stream_lines_output) if stream_lines_output else ""
-        offset_lines_output = list(display.offset_lines)
-        audio_block["offset_lines"] = offset_lines_output
-        audio_block["offset_lines_text"] = "\n".join(offset_lines_output) if offset_lines_output else ""
-        measurement_source = dict(display.measurements)
-        if (
-            not measurement_source
-            and summary is not None
-            and summary.measured_offsets
-        ):
-            measurement_source = {
-                detail.label: detail for detail in summary.measured_offsets.values()
-            }
-        measurements_output: dict[str, dict[str, object]] = {}
-        for label, detail in measurement_source.items():
-            measurements_output[label] = {
-                "stream": detail.stream,
-                "seconds": detail.offset_seconds,
-                "frames": detail.frames,
-                "correlation": detail.correlation,
-                "status": detail.status,
-                "applied": detail.applied,
-                "note": detail.note,
-            }
-        audio_block["measurements"] = measurements_output
-        if display.manual_trim_lines:
-            audio_block["manual_trim_summary"] = list(display.manual_trim_lines)
-        else:
-            audio_block["manual_trim_summary"] = []
-        if display.warnings and collected_warnings is not None:
-            collected_warnings.extend(display.warnings)
-    else:
-        audio_block["reference_stream"] = None
-        audio_block["target_stream"] = cast(dict[str, object], {})
-        audio_block["offsets_sec"] = cast(dict[str, object], {})
-        audio_block["offsets_frames"] = cast(dict[str, object], {})
-        audio_block["manual_trim_summary"] = []
-        audio_block["stream_lines"] = []
-        audio_block["stream_lines_text"] = ""
-        audio_block["offset_lines"] = []
-        audio_block["offset_lines_text"] = ""
-        audio_block["measurements"] = {}
-
-    audio_block["enabled"] = bool(cfg.audio_alignment.enable)
-    audio_block["suggestion_mode"] = bool(summary.suggestion_mode if summary else False)
-    audio_block["suggested_frames"] = dict(summary.suggested_frames) if summary else {}
-    audio_block["manual_trim_starts"] = dict(summary.manual_trim_starts) if summary else {}
-    audio_block["vspreview_manual_offsets"] = (
-        dict(summary.vspreview_manual_offsets) if summary else {}
-    )
-    audio_block["vspreview_manual_deltas"] = (
-        dict(summary.vspreview_manual_deltas) if summary else {}
-    )
-    if (
-        summary is not None
-        and summary.reference_plan.path.name in summary.vspreview_manual_offsets
-    ):
-        audio_block["vspreview_reference_trim"] = int(
-            summary.vspreview_manual_offsets[summary.reference_plan.path.name]
-        )
+def _safe_float(value: object) -> float | None:
+    if isinstance(value, (int, float)):
+        float_value = float(value)
+        if math.isnan(float_value):
+            return None
+        return float_value
+    return None
 
 
-def apply_manual_offsets_logic(
-    plans: Sequence[ClipPlan],
-    vspreview_reuse: dict[str, int],
-    display_data: _AudioAlignmentDisplayData,
-    plan_labels: dict[Path, str],
-) -> tuple[dict[str, int], dict[str, int]]:
-    """
-    Apply manual offsets from VSPreview history, normalizing to avoid negative trims (padding).
-    Extracted for testability.
-
-    Returns:
-        tuple[dict[str, int], dict[str, int]]: (delta_map, manual_trim_starts)
-    """
-    if display_data.manual_trim_lines:
-        display_data.manual_trim_lines.clear()
-
-    baseline_map = {plan.path.name: int(plan.trim_start) for plan in plans}
-    delta_map: dict[str, int] = {}
-    manual_trim_starts: dict[str, int] = {}
-
-    # 1. Calculate proposed trims for all plans
-    proposed_trims: dict[str, int] = {}
-    for plan in plans:
-        key = plan.path.name
-        baseline = baseline_map.get(key, int(plan.trim_start))
-        delta = int(vspreview_reuse.get(key, 0))
-        proposed_trims[key] = baseline + delta
-        if key in vspreview_reuse:
-            delta_map[key] = delta
-
-    logger.info(f"[ALIGN DEBUG] Raw manual offsets (vspreview_reuse): {vspreview_reuse}")
-    logger.info(f"[ALIGN DEBUG] Baseline trims: {baseline_map}")
-    logger.info(f"[ALIGN DEBUG] Proposed trims before normalization: {proposed_trims}")
-
-    # 2. Normalize to ensure no negative trims (avoid padding)
-    # Normalize by shifting all clips up so the minimum trim becomes 0
-    # Example: offsets [-361, 0] → shift +361 → trims [0, 361]
-    min_offset = min(proposed_trims.values()) if proposed_trims else 0
-    shift = -min_offset if min_offset < 0 else 0
-
-    if shift > 0:
-        logger.info(f"[ALIGN DEBUG] Found negative proposed trim {min_offset}, applying global shift of {shift}f.")
-    else:
-        logger.info("[ALIGN DEBUG] No negative proposed trims found, no global shift applied.")
-
-    logger.info(f"[ALIGN DEBUG] Normalizing with shift={shift}")
-
-    # 3. Apply normalized trims
-    for plan in plans:
-        key = plan.path.name
-        if key in proposed_trims:
-            raw_offset = proposed_trims[key]
-            # Calculate normalized trim
-            normalized_trim = raw_offset + shift
-
-            # Apply to plan
-            # Note: We SET the new absolute trim_start (normalized_trim includes the baseline)
-            original_trim = int(plan.trim_start)
-            if normalized_trim != original_trim:
-                plan.trim_start = normalized_trim
-                plan.source_num_frames = None
-                plan.has_trim_start_override = True
-
-            manual_trim_starts[key] = int(plan.trim_start)
-
-
-            baseline = baseline_map.get(key, 0)
-            effective_delta = normalized_trim - baseline
-            delta_map[key] = effective_delta
-
-            label = plan_labels.get(plan.path, key)
-            input_delta = proposed_trims[key] - baseline
-
-            line = (
-                f"VSPreview manual offset applied: {label} baseline {baseline}f "
-                f"{input_delta:+d}f (shift {shift:+d}f) → {normalized_trim}f"
-            )
-            display_data.manual_trim_lines.append(line)
-
-            logger.info(f"[ALIGN DEBUG] Applied trim to {key}: raw={raw_offset} + shift={shift} -> {normalized_trim}. Final trim_start={plan.trim_start}")
-
-
-
-    return delta_map, manual_trim_starts
+def _safe_int(value: object) -> int | None:
+    if isinstance(value, (int, float)):
+        float_value = float(value)
+        if math.isnan(float_value):
+            return None
+        return int(float_value)
+    return None
